Conflicts:
	src/gui/moviebrowser/mb.cpp


Origin commit data
------------------
Commit: https://github.com/TangoCash/neutrino-tangos/commit/c387d1ce740917631d1159d5b76fc18d9ba2739e
Author: TangoCash <eric@loxat.de>
Date: 2016-09-30 (Fri, 30 Sep 2016)



------------------
This commit was generated by Migit

+++ src/gui/movieplayer.cpp
@@ -1664,7 +1664,7 @@ void CMoviePlayerGui::PlayFileLoop(void)
 		} else if (msg == (neutrino_msg_t) g_settings.key_quickzap_up || msg == (neutrino_msg_t) g_settings.key_quickzap_down) {
 			quickZap(msg);
 		} else if (fromInfoviewer && msg == CRCInput::RC_ok && !filelist.empty()) {
-			printf("CMoviePlayerGui::%s: start playlist movie #%d\n", __func__, vzap_it - filelist.begin());
+			printf("CMoviePlayerGui::%s: start playlist movie #%d\n", __func__, (int)(vzap_it - filelist.begin()));
 			fromInfoviewer = false;
 			playstate = CMoviePlayerGui::STOPPED;
 			filelist_it = vzap_it;
@@ -2026,7 +2026,7 @@ void CMoviePlayerGui::PlayFileEnd(bool restore)
 
 void CMoviePlayerGui::set_vzap_it(bool up)
 {
-	//printf("CMoviePlayerGui::%s: vzap_it: %d count %s\n", __func__, vzap_it - filelist.begin(), up ? "up" : "down");
+	//printf("CMoviePlayerGui::%s: vzap_it: %d count %s\n", __func__, (int)(vzap_it - filelist.begin()), up ? "up" : "down");
 	if (up)
 	{
 		if (vzap_it < (filelist.end() - 1))
@@ -2037,7 +2037,7 @@ void CMoviePlayerGui::set_vzap_it(bool up)
 		if (vzap_it > filelist.begin())
 			--vzap_it;
 	}
-	//printf("CMoviePlayerGui::%s: vzap_it: %d\n", __func__, vzap_it - filelist.begin());
+	//printf("CMoviePlayerGui::%s: vzap_it: %d\n", __func__, (int)(vzap_it - filelist.begin()));
 }
 
 void CMoviePlayerGui::callInfoViewer(bool init_vzap_it)
